Move adapters into internal package to highlight the intent to restrict their use.
