+++ webserver/cors/src/main/java/io/helidon/webserver/cors/CORSSupport.java
@@ -25,6 +25,8 @@ import io.helidon.webserver.Routing;
 import io.helidon.webserver.ServerRequest;
 import io.helidon.webserver.ServerResponse;
 import io.helidon.webserver.Service;
+import io.helidon.webserver.cors.internal.RequestAdapter;
+import io.helidon.webserver.cors.internal.ResponseAdapter;
 
 /**
  * A Helidon service and handler implementation that implements CORS, for both the application and for built-in Helidon
