+++ webserver/cors/src/main/java/io/helidon/webserver/cors/SEResponseAdapter.java
@@ -18,6 +18,7 @@ package io.helidon.webserver.cors;
 
 import io.helidon.common.http.Http;
 import io.helidon.webserver.ServerResponse;
+import io.helidon.webserver.cors.internal.ResponseAdapter;
 
 /**
  * SE implementation of {@link ResponseAdapter}.
